+++ src/RequestHandlerInterface.h
@@ -14,7 +14,7 @@ protected:
   virtual void showTopicPartitionOffsets(UserArgumentStruct UserArguments) = 0;
 
   virtual void subscribeConsumeAtOffset(std::string TopicName,
-                                        int32_t Offset) = 0;
+                                        int64_t Offset) = 0;
   virtual void subscribeConsumeNLastMessages(std::string TopicName,
                                              int64_t NumberOfMessages,
                                              int Partition) = 0;
